requirements.txt
discord.py
python-dotenv
gspread
google-auth
python-dateutil
pytz